fix: move lib to lib and build mock
